// payment_request/models/PaymentMethodType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NOTE: This file is auto generated by Xendit.
 * Do not edit the class manually.
 * Improvements? Share your ideas at https://github.com/xendit/xendit-node
 */


/**
 * 
 * @export
 */
export const PaymentMethodType = {
    Card: 'CARD',
    DirectDebit: 'DIRECT_DEBIT',
    Ewallet: 'EWALLET',
    OverTheCounter: 'OVER_THE_COUNTER',
    QrCode: 'QR_CODE',
    VirtualAccount: 'VIRTUAL_ACCOUNT',
    XenditEnumDefaultFallback: "UNKNOWN_ENUM_VALUE"
} as const;
export type PaymentMethodType = typeof PaymentMethodType[keyof typeof PaymentMethodType];


export function PaymentMethodTypeFromJSON(json: any): PaymentMethodType {
    return PaymentMethodTypeFromJSONTyped(json, false);
}

export function PaymentMethodTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): PaymentMethodType {
    if (json !== "" || json !== null) {
        const key = Object.keys(PaymentMethodType)[Object.values(PaymentMethodType).indexOf(json)]
        return PaymentMethodType[key] === undefined ?
            PaymentMethodType['XenditEnumDefaultFallback'] : PaymentMethodType[key]
    }
    return json as PaymentMethodType;
}

export function PaymentMethodTypeToJSON(value?: PaymentMethodType | null): any {
    return value as any;
}